REF: Move asarray_multi functions to base Operator
These functions do not need to be GPU specific
